Made home page imports dynamic
