+++ .vscode/settings.json
@@ -1,8 +1,5 @@
 {
   "editor.tabSize": 2,
   "editor.formatOnSave": true,
-  "editor.codeActionsOnSave": {
-    "source.organizeImports": true,
-  },
   "WillLuke.nextjs.hasPrompted": true
 }
