+++ package.json
@@ -25,11 +25,10 @@
     "@prisma/client": "5",
     "@tabler/icons": "^1.106.0",
     "@tanstack/react-query": "^4.10.0",
-    "@trpc/client": "^10.40.0",
-    "@trpc/next": "10.0.0-proxy-beta.17",
-    "@trpc/react": "10.0.0-proxy-beta.17",
-    "@trpc/react-query": "^10.40.0",
-    "@trpc/server": "^10.40.0",
+    "@trpc/client": "^10.41.0",
+    "@trpc/next": "^10.41.0",
+    "@trpc/react-query": "^10.41.0",
+    "@trpc/server": "^10.41.0",
     "cookies-next": "^2.1.1",
     "cuid": "^2.1.8",
     "dayjs": "^1.11.6",
@@ -37,7 +36,7 @@
     "next-auth": "^4.12.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "recharts": "^2.4.1",
+    "recharts": "^2.9.0",
     "superjson": "1.9.1",
     "zod": "^3.18.0"
   },
@@ -46,14 +45,14 @@
     "@types/bcryptjs": "^2.4.2",
     "@types/node": "18.0.0",
     "@types/nodemailer": "^6.4.6",
-    "@types/react": "^18.0.14",
-    "@types/react-dom": "^18.0.5",
+    "@types/react": "^18.2.29",
+    "@types/react-dom": "^18.2.14",
     "@typescript-eslint/eslint-plugin": "^5.33.0",
     "@typescript-eslint/parser": "^5.33.0",
     "JSONStream": "^1.3.5",
     "autoprefixer": "^10.4.16",
     "bcryptjs": "^2.4.3",
-    "eslint": "8.22.0",
+    "eslint": "^8.51.0",
     "eslint-config-next": "^13.5.5",
     "eslint-config-prettier": "^9.0.0",
     "nodemailer": "^6.8.0",
@@ -61,9 +60,10 @@
     "prettier": "^3.0.3",
     "prettier-plugin-tailwindcss": "^0.5.6",
     "prisma": "5",
+    "prop-types": "^15.8.1",
     "tailwindcss": "^3.3.3",
     "ts-node": "^10.9.1",
-    "typescript": "^5.1.3"
+    "typescript": "^5.2.2"
   },
   "ct3aMetadata": {
     "initVersion": "6.5.0"
